hal9003/requirements.txt
autobahn
open-chat-api-client @ https://github.com/tbscode/open-chat-api-client/tarball/master
openai
redis
watchdog
py-mon
django
djangorestframework
aiohttp
requests
jsonschema
tabulate
pyyaml